+++ lib/plugins/method/__init__.py
@@ -1,0 +1,15 @@
+from yapsy.IPlugin import IPlugin
+from lib.methods import BaseMethod
+
+class IMethodPlugin(BaseMethod, IPlugin):
+
+  def __init__(self):
+    pass
+
+  def setNameAndFactory(self, name, factory):
+    self.methodName = name
+    self.factory = factory
+
+  @staticmethod
+  def supports(methodName):
+    raise NotImplementedError