+++ lib/plugins/task/__init__.py
@@ -1,0 +1,6 @@
+from yapsy.IPlugin import IPlugin
+from fabric.tasks import Task
+
+class ITaskPlugin(Task, IPlugin):
+  def run(self):
+    raise NotImplementedError