+++ Changelog.md
@@ -3,6 +3,47 @@
 ## 2.x.x
 
 ### new
+
+* new section in fabfile.yaml called `blueprints`. THis section allows the usage of a given sets of blueprints as normal config. That means, they will get listed and work wirh autocomplete
+
+      ```
+      blueprints:
+        - configName: <name of blueprint-config>
+          variants:
+            - variantA
+            - variantB
+            - variantC
+      ```
+
+      The list of configs will include the three variants applied with the blueprint `<name of blueprint-config>` This will allow a config inherited from a blueprint with a given variant.
+
+
+* new plugin-discovery via `yapsy`, it allows you to add custom tasks and methods to your project, or to your user-folder.
+
+    The plugin-resolver will look into the following folders to find tasks and methods:
+
+      ```
+      <project_folder>/.fabalicious/plugins
+      <user-folder>/.fabalicious/plugins
+      <fabalicious-folder>/tasks
+      ```
+
+### changed
+
+* methods are now responsible to provide their default global settings via the static `getGlobalSettings`-method.
+
+## 2.3.1
+
+### new
+* new task `logLevel`: used to change the default logging level of the program.
+
+### changed
+
+* Refactored output logging to use python logging library.
+
+## 2.3.0
+
+### new
 * new task `sshCommand`: it will print out a ssh command to log into a remote instance. With some shell-magic you can streamline it into a single shell-command. This command will not have the downsides of the `ssh`-command e.g. the lagging shell.
 * new task `installFrom` which will do a install-task and a copyFrom afterwards. It will run the reset-task only after the copyFrom, so performance is better
 * new task `getFilesDump` which will tar all files in the `filesFolder` into a tar and download it to the local computer.
@@ -214,5 +255,3 @@
 * `needsComposer` is unsupported, set your `needs` accordingly.
 * the custom script-command `run_task` is not supported anymore. Use `execute(<task-name>)` instead.
 * the task `updateDrupalCore` is not ported over, not sure if it comes back.
-
-